Make unused(*) modules private
There are two technical exceptions here. pyiron_workflow.wf_graph is imported in the pyiron_core_dev_full.ipynb notebook, but that was last touched in the repo over three months ago, so I'm simply assuming it's not longer particularly relevant. The second is that pyiron_workflow.graph.graph_operations _is_ imported in the tests.ipynb notebook, which was updated only last week and is presumably still very relevant, _but_ the import is not used at all.