Fix storage
Even in the absence of a database, we want to support storage. Changes from refactored_jsx added a clause for storing the database, which internally calls `idb.store_node_in_database` when `store_outputs=True`, but commented out the storage for the non-database case. This is re-introduced behind an `else`.