Localize circular import
Closes https://github.com/pyiron/pyiron_core/issues/84.

IMO this is smelly. The GUI functionality should not be referenced in the serialization back-end. The private function should not be referenced by an external module. Nonetheless, let's leave the solution this way so that at least the interaction of collapsed/expanded with the serialization back end has a clear trail.