Fix the IterNode node to work with the new db setup
The new db infrastructure isn't propagated through the entire codebase -- the simple_workflow level knows about it but the graph level largely doesn't -- so we need to hard-fail if there's an executor _and_ a db. Also, because of the way IterNode on-the-fly invokes the run method of other nodes, we still wind up needing the database as an explicit input, which goes against the spirit of the new infrastructure. But, it gets back the desired results and is robust (i.e. works when it can, fails cleanly when it can't) so we'll live with it.