Don't give paths to non-existant functions (#122)
* Pyironflow api (#110)

* Create api for pyironflow submodule

Based on what is used from outside the module

* Leverage pyironflow api

* Pyiron database api (#112)

* Create api for pyiron_database submodule

Based on what is used from outside the module

* Leverage pyiron_database api

* Pyiron nodes inits (#113)

* Alphabetize and clean __init__.py

* Add __init__.py files to all submodules

* Add remaining node files/modules to import

* Expose pyiron_nodes as nodes (#114)

And demonstrate usage in user-facing notebook examples. Internally, we stick with referencing the module itself.

* Top-level user API (#115)

* Expose key user-facing pyiron_workflow stuff

Right in the pyiron_core init

* Use top-level imports in demo notebooks

* Tidying up (#116)

* Expose util in the top api

* Remove unused import

* Remove commented code

* Fix indent

* Reorder imports

* Re-execute notebooks

* Copy workflows from gitlab

* Re-path json files

Based on refactoring into a single python package.

* Add landau dependency

* Explicitly allow integer vacancy indices

Otherwise, the "grace_minimize" demo refuses to load because you can't use `in` on an integer. Indeed, I don't understand how this node would run in isolation without this fix.

* Give a GUI-level flag for loading compact

The whole flag is only needed until the legacy workflows can get upgraded (https://github.com/pyiron/pyiron_core/issues/91) (or the legacy workflows are simply _dropped_). I think it would be even better to switch to the "compact" state being the default expectation, but that can wait until a future PR.

* Start testing demos

To begin, ensure they all load without error

* Add assyst

* Fix whitespace

* Test running the landau example

* Bump pyiron_atomistics

* Add structuretoolkit

There is an explicit dependence in pyiron_nodes

* Remove tensorpotential

As far as I can tell, what I actually need is grace-tensorpotential.

* Start isolating architecture-dependent dependencies

For at least running the workshop demo CI remotely

* Download data per-test

Instead of universally in the test class

* remove unused variable

* Use the optional dependencies in CI

* Test linearfit if dependencies are available

I snagged the reference output from the mpcdf binder notebook. I'm unable to run locally, so let's see if the test is good...

* :bug: move env file to the right dir

* Refactor data generation

* Fix package name

* Get data

* Use consistent indentation

* Revert print statements

from debugging gui impersonation

* Expose phonons submodule

* Test phonons workflow

* Amend job name

* Increase timeout window

* Extend dependencies

Based on workshop binder env https://gitlab.mpcdf.mpg.de/pyiron/pyiron_core_tutorial/-/blob/b55b388adc8a407b6344eff220beefccfefead63/binder/environment.yml.

I already took the liberty of using more recent versions in some cases, so this may need to be regressed to match the workshop repo. Additionally, I prefer to only include explicit dependencies, and some are missing. In a "trust but verify" approach, I first want to get the tests running on the remote machine and then go back and remove what appear to be superfluous dependencies to verify that they're present for a good reason.

Not found but I already know there's a good reason:
- iprpy-data
  - this is probably indirectly referenced by pyiron_*
- nglview
  - this is an optional dependency of some of our atomistics code, so we need it explicitly here to implicitly leverage it there
- pyxtal
  - An optional dependency in structuretoolkit, made explicit here; cf. @pmrv's comment in [the assyst setup](https://github.com/eisenforschung/assyst/blob/254c93171d5044bc5398c46f1ca0f76ff6ae4ad7/pyproject.toml#L29-L32) for why this is a good idea

Suspect:
- bayesian-optimization `bayes_opt`
- python-ternary `ternary`
- scikit-image
- scikit-learn

* Organize dependencies

* Test the water workflow

I'm not sure what a reasonable p-threshold is yet

* Encode the similarity test

* See what grace has for us

* Encode grace results

It's an actual minimization, not a stochastic MD, so hopefully we are safe to target the actual end result quite precisely.

* Test energy workflow

* Use unittest tool

I forgot the name of it until now, but for simple float comparison let's use it directly

* Put a placeholder for the elastic test

* Test calphy

Against the entire output energy

* Check if the calphy run is the problem

* Add calphy back

* Update calphy test

- Remove prints
- Detect and fail cleanly on a melt (what else to do??)
- Relax comparison -- this is stochastic

* Run calphy at a lower temperature

* Remove loading test

They each get loaded in their own test

* Actually run it

* Remove calphy2 workflow

It has been unloadable forever, and the new demos contain a calphy example

* Replace unimportable function

With the serial api reference to an unimportable function

* Update docstring

* Use direct import

* Revert re-pathing

in graph.group; we'll do it directly when getting the import path.

* Deduplicate function

get_import_path_from_type should not exist in two places with such similar functionality. We'll deal with the non-top import later

* Flag and catch code from graphs

* Get name from types safely

* Only `serial`-ize `serial` objects

* Remove explicit None handling

It was faulty -- we want the path to the None object for simple_workflow usage. But anyhow, now that we grab the module and name variables first, it's not necessary

* Test new behaviour