Add import for assyst nodes
Both the assyst and assyst 2 packages depend on this. Clearly, these nodes aren't showing up in the tests anywhere. The conda package doesn't support osx-arm64, so I'm not going to spend time trying to test these nodes (plus there's two versions of the package anyway...) but we can at least be clear about the imports and requirements.