Add ruff to the CI
This increases the immediate burden on developers making a PR as they need to ensure that work stays compatible with ruff formatting in order to pass the CI suite; however, in the long run it is considered a best-practice for reducing the overall burden on developers as it becomes easier to read and parse the code base. It is common [but not universal](https://github.com/pyiron/decisions/issues/5) in pyiron repositories.