Symmetrically update when adding an edge
`self.update_gui()` and `self.main_widget.redraw()` are quite similar. Experimenting with this change, I didn't see a functional difference, but `update_gui` is snappier and less visually disruptive (no flicker). Additionally, this approach gives better alignment between the different branches of the if clause here, the rest of which use `update_gui` (if anything).